feat: move rag pipeline to different file - add fastapi and post endpoint
